Fix for circular reference error
On create_subnet, the quantum client passes in arguments to create a subnet
with, including dns_nameservers and allocation_pools, which do not exist in
quark's database models. SQLAlchemy causes these fields to exist in the
model created and result as basic objects in the response object, which in turn
is choked upon by oslo-incubator's jsonutils.to_primitive function.