Implements a first cut of network strategies
Implements the basic network strategy driver. The aim of this patch is
to provide a conceptually similar implement of both provider networks
and melange default networks while adding additional flexibility. One of
the other aims of this patch is to attempt to coerce a truer version of
shared networks out of the quantum implementation as it exists today.

Still considered a work in progress.