Allows easier configuration of default networks
Updates the Quark networks extension to allow the id and ipam_strategy
to be passed along with a network create. The id is specifically gated
as admin only, but as an admin, allows deployers to configure special
networks like the all 0s and 1s placeholder values we use at Rackspace.