Fixes #29 and #31
Now honors "do_not_use" on subnets, effectively ignoring them. Also, as
per issue 31, max_allocation is honored correctly so long as we 1) know
where to port max_allocation to (the quotas table) and 2) always create
the quotas table. The way to do that is to import the quota db driver in
the models regardless of whether we're using the conf driver or the db
driver. This is seemingly the blessed method upstream, so we'll follow
suit.