Fix non-deterministic error in unit test
next_auto_assign_ip is set in the cidr's constructor so depending on
the subnet dictionary key memory layout next_auto_assign_ip would be
set to 0 as opposed to 2 if the cidr's constructor was invoked after
next_auto_assign_ip was manually set. This was fixed by updating
next_auto_assign_ip after the subnet model was created.