Default CIDRs in IPPolicy related to Allocation Pools
Via help from @asadoughi, was able to figure out a way to get an IPV6 subnet
into a netaddr.IPRange.  That allowed us to set allocation pools for IPV6
subnets on create.

Also fixed notification test.  Failed when run individually, passed when
run with test suite.  Mocked out _validate_subnet_cidr as it was expecting
a non-single scopt of subnet_find, but delete_subnet is expecting a single
scope of subnet_find

RM4900