Fixes #118 again
Seemingly, the pop(...) used in the plugin_view call was disassocating
the route from the subnet when it was listed, causing a very subtle but
heinous bug. This corrects that by instead conditionally adding routes
to the view rather than popping them from the model before adding them
all.

Additionally, corrects an assumption made earlier that creating multiple
default routes actually failed, when it in fact does not.