Fixes RM7946
Now that we have alembic migrations, the need for quark to automatically
create the schema is no longer required. This commit
removes two lines that cause sqlalchemy to create the schema in the
database if one does not exist.

Additionally, quite a few port and subnet tests were accidentally
leaking past the test abstraction and hitting the database. This patch
unearthed that problem, so several tests were modified to fix the issue.