Made it so shared=false works properly in noauth
RM8126

noauth was making it so shared=false + no tenant filter was returning shared
networks even though shared was supposed to be filtered out. The reason why
a sentinel value (INVERT_DEFAULTS) is used instead of an additional parameter
was due to it being the path of least resistance. There are intentions to
eventually move the shared attribute to the model.