Fix PUT to port with fixed ips
RM9097

Ran into an unrelated problem which needed to be fixed: device_id is not
required by neutron and shouldn't have been. quark will now set the device_id
to empty string should the device_id not exist in the port request body.

This fixes the ipv6->ipv4 conversion error when updating ports.