Fixes update_port to handle empty security_groups
RM10261

Updates the handling of security groups in update_port to allow no
security groups to be passed at all. Previously, if groups were not
passed, that was interpreted as the same thing as passing an empty list
of groups, mistakenly calling the redis portion of the update handling,
and could have potentially wiped out flows for any port that had
security groups.