Removes superfluous pessimistic connection code
Removing RM9078

Updates the quark db layer to remove pessmistic connection handling,
added by RM9078. The code was put in place to provide assistance around a
perceived bug at the time, but was never utilized and was also not the
source of the issue. Most importantly, oslo.db already provides
identical functionality, so the original patch was a waste.