redis_sg_tool now installs + fixes
RM11081

Moves redis_sg_tool to the quark namespace and adds an entry point to
setup.cfg for it. Additionally removes a group_by from
ports_with_security_groups_count db_api call that was breaking the
ports-with-groups command line action.

Finally, a bunch of PEP8 fixes because it was ignored by tox before.