Makes the agent apply flows to the correct VIF
RM11249

Previously, the agent assumed that the order of VIFs returned from Xen
was significant, and used the index in the list to called the flows
plugin. However, that isn't always strictly true. The right answer is to
get the device number from the VIF record itself and pass that into the
call.