Replaces agent JSON control with Redis control
RM11449

This patch rips out the JSON version control system implemented in a
previous version of the agent and replaces it with a new hash field in
Redis. The agent had to be substantially reworked to compensate for the
change. Additionally, this patch enables duplex communication for a
process using the quark redis client by implementing discreet read and
write only connection pools.