Raise different errors if it's a provider network
JIRA:NCP-1660

IPAM logic will now raise a different error for provider networks than
for tenant networks. The reasoning is the failure to have enough IP
addresses for a provider network is a provider problem, and therefore
should be a 500 error. Meanwhile, tenant network IP address failures
remain a 409, as that particular failure is still most likely on the
customer and their usage.