Fix SegmentAllocationRange queries for non-elevated context
Add SegmentAllocationRange to the list of models that do not
filter on tenant_id to avoid blowing up in _model_query.

This is queried when a network is created, so non-admin users need to be
able to read this table, which means tenant_id is injected into the
query. However, the SegmentAllocationRange model doesn't have a tenant_id,
so we need to avoid filtering on that.

JIRA:NCP-1804