Fix 500 on missing "cidr" key from POST to subnets
If cidr is missing from request body on a POST to subnets,
neutron will throw a 500. Quark uses netaddr, so the Sentinel
passed in for a missing attribute cannot be handled properly.
Wrap the netaddr.IPSet create call in a try/catch to check
for TypeError and throw a BadRequest.

Change-Id: Iee4d874551b859db8239ba6fe9743662755198dc
JIRA:NCP-1990
Closes-Bug: #1598956