Prevent flake8 errors from getting suppressed
Flake8 errors, if present, do not show up if run_tests.sh is used. It is
used by Jenkins on creation of a pull request, upon which the tests fail
without flake8 specific information showing up in the log.
Adding PYTHONUNBUFFERED=1 before the flake8 command in run_tests.sh solves
this problem.

Change-Id: I040873b859cca7c0500bd7d600d8de6121cc70a5
Closes-Bug: #1747420